fix the code and split the file
