+++ src/components/Recruiter-Pipeline/recruiter-pipeline.tsx
@@ -7,211 +7,14 @@ import { Input } from "@/components/ui/input";
 import { KPISection } from "./kpi-section";
 import { CreatePipelineDialog } from "./create-pipeline-dialog";
 import { PipelineJobCard } from "./pipeline-job-card";
-import { type Job, mapUIStageToBackendStage, mapBackendStageToUIStage } from "./dummy-data";
+import { type Job, mapUIStageToBackendStage } from "./dummy-data";
+import { convertPipelineListDataToJob, convertPipelineDataToJob } from "./utils/convert";
 import { Button } from "../ui/button";
 import { toast } from "sonner";
 import { useAuth } from "@/contexts/AuthContext";
 import { getAllPipelineEntries, getPipelineEntry, updateCandidateStage as updateCandidateStageAPI, type PipelineListItem } from "@/services/recruitmentPipelineService";
 
-// Utility function to convert pipeline list data to Job format
-const convertPipelineListDataToJob = (pipelineData: PipelineListItem, isExpanded: boolean = false): Job => {
-  // Format salary range using the correct backend fields
-  const formatSalaryRange = () => {
-    const minSalary = pipelineData.jobId.minimumSalary;
-    const maxSalary = pipelineData.jobId.maximumSalary;
-    const currency = pipelineData.jobId.salaryCurrency;
-    
-    if (minSalary && maxSalary && currency) {
-      return `${minSalary} - ${maxSalary} ${currency}`;
-    } else if (minSalary && currency) {
-      return `${minSalary} ${currency}`;
-    } else if (maxSalary && currency) {
-      return `${maxSalary} ${currency}`;
-    } else if (minSalary || maxSalary) {
-      return `${minSalary || maxSalary}`;
-    }
-    return "Salary not specified";
-  };
-
-  return {
-    id: pipelineData._id,
-    title: pipelineData.jobId.jobTitle || "Untitled Job",
-    clientName: pipelineData.jobId.clientName || "Unknown Client",
-    location: pipelineData.jobId.location || "Location not specified",
-    salaryRange: formatSalaryRange(),
-    headcount: pipelineData.jobId.numberOfPositions || 1,
-    jobType: pipelineData.jobId.jobType || "Full-time",
-    isExpanded,
-    // Preserve the entire jobId object for access to jobTeamInfo
-    jobId: pipelineData.jobId,
-    candidates: [], // Will be populated when expanded
-    // Pipeline-specific data from new API structure
-    // Note: pipelineStatus is now derived from jobId.stage
-    priority: pipelineData.priority,
-    notes: pipelineData.notes,
-    assignedDate: pipelineData.assignedDate,
-    // Candidate counts from new API structure
-    totalCandidates: pipelineData.totalCandidates,
-    activeCandidates: pipelineData.activeCandidates,
-    completedCandidates: pipelineData.completedCandidates,
-    droppedCandidates: pipelineData.droppedCandidates,
-    numberOfCandidates: pipelineData.numberOfCandidates,
-    // Recruiter information
-    recruiterName: pipelineData.recruiterId?.name || "Unknown Recruiter",
-    recruiterEmail: pipelineData.recruiterId?.email || "",
-    // Job details from API
-    department: pipelineData.jobId.department,
-    numberOfPositions: pipelineData.jobId.numberOfPositions,
-    // Additional fields will be populated when detailed data is loaded
-  };
-};
-
-// Utility function to convert detailed pipeline data to Job format
-const convertPipelineDataToJob = (pipelineData: any, isExpanded: boolean = false): Job => {
-  // Format salary range using the correct backend fields
-  const formatSalaryRange = () => {
-    const minSalary = pipelineData.jobId.minimumSalary;
-    const maxSalary = pipelineData.jobId.maximumSalary;
-    const currency = pipelineData.jobId.salaryCurrency;
-    
-    if (minSalary && maxSalary && currency) {
-      return `${minSalary} - ${maxSalary} ${currency}`;
-    } else if (minSalary && currency) {
-      return `${minSalary} ${currency}`;
-    } else if (maxSalary && currency) {
-      return `${maxSalary} ${currency}`;
-    } else if (minSalary || maxSalary) {
-      return `${minSalary || maxSalary}`;
-    }
-    return "Salary not specified";
-  };
-
-  return {
-    id: pipelineData._id,
-    title: pipelineData.jobId.jobTitle || "Untitled Job",
-    clientName: pipelineData.jobId.client?.name || "Unknown Client",
-    location: pipelineData.jobId.location || "Location not specified",
-    salaryRange: formatSalaryRange(),
-    headcount: pipelineData.jobId.headcount || 1,
-    jobType: pipelineData.jobId.jobType || "Full-time",
-    isExpanded,
-    // Preserve the entire jobId object for access to jobTeamInfo
-    jobId: pipelineData.jobId,
-    candidates: pipelineData.candidateIdArray.map((candidateData: any) => {
-      const candidate = candidateData.candidateId;
-      return {
-        id: candidate._id,
-        name: candidate.name,
-        source: candidate.referredBy || "Pipeline",
-        currentStage: mapBackendStageToUIStage(candidateData.currentStage || "Sourcing"),
-        avatar: undefined,
-        experience: candidate.experience,
-        currentSalary: candidate.currentSalary,
-        currentSalaryCurrency: candidate.currentSalaryCurrency,
-        expectedSalary: candidate.expectedSalary,
-        expectedSalaryCurrency: candidate.expectedSalaryCurrency,
-        currentJobTitle: candidate.currentJobTitle,
-        previousCompanyName: candidate.previousCompanyName,
-        // Additional fields from new structure
-        applicationId: candidate._id,
-        appliedDate: candidateData.addedToPipelineDate,
-        lastUpdated: candidateData.lastUpdated,
-        applicationDuration: 0, // Calculate if needed
-        // Candidate details
-        email: candidate.email,
-        phone: candidate.phone,
-        location: candidate.location,
-        skills: candidate.skills,
-        softSkill: candidate.softSkill,
-        technicalSkill: candidate.technicalSkill,
-        gender: candidate.gender,
-        dateOfBirth: candidate.dateOfBirth,
-        country: candidate.country,
-        nationality: candidate.nationality,
-        willingToRelocate: candidate.willingToRelocate,
-        description: candidate.description,
-        linkedin: candidate.linkedin,
-        reportingTo: candidate.reportingTo,
-        // Additional fields for dialog
-        educationDegree: candidate.educationDegree,
-        primaryLanguage: candidate.primaryLanguage,
-        resume: candidate.resume,
-        // Pipeline-specific data
-        status: candidateData.status,
-        subStatus: candidateData.status,
-        priority: candidateData.priority,
-        notes: candidateData.notes,
-        // Stage-specific data - IMPORTANT: Include all stage data
-        sourcing: candidateData.sourcing,
-        screening: candidateData.screening,
-        clientScreening: candidateData.clientScreening,
-        interview: candidateData.interview,
-        verification: candidateData.verification,
-        onboarding: candidateData.onboarding,
-        hired: candidateData.hired,
-        disqualified: candidateData.disqualified,
-        // Additional pipeline fields
-        connection: candidateData.connection,
-        hiringManager: candidateData.hiringManager,
-        recruiter: candidateData.recruiter,
-        isTempCandidate: candidate.isTempCandidate || false,
-      };
-    }),
-    // Pipeline-specific data from new API structure
-    // Note: pipelineStatus is now derived from jobId.stage
-    priority: pipelineData.priority,
-    notes: pipelineData.notes,
-    assignedDate: pipelineData.assignedDate,
-    // Candidate counts from new API structure
-    totalCandidates: pipelineData.totalCandidates,
-    activeCandidates: pipelineData.activeCandidates,
-    completedCandidates: pipelineData.completedCandidates,
-    droppedCandidates: pipelineData.droppedCandidates,
-    numberOfCandidates: pipelineData.totalCandidates,
-    // Recruiter information
-    recruiterName: pipelineData.recruiterId?.name || "Unknown Recruiter",
-    recruiterEmail: pipelineData.recruiterId?.email || "",
-    // Job details from API
-    jobPosition: Array.isArray(pipelineData.jobId.jobPosition) ? 
-      pipelineData.jobId.jobPosition.join(", ") : 
-      pipelineData.jobId.jobPosition,
-    department: pipelineData.jobId.department,
-    experience: pipelineData.jobId.experience,
-    education: Array.isArray(pipelineData.jobId.education) ? 
-      pipelineData.jobId.education.join(", ") : 
-      pipelineData.jobId.education,
-    specialization: Array.isArray(pipelineData.jobId.specialization) ? 
-      pipelineData.jobId.specialization.join(", ") : 
-      pipelineData.jobId.specialization,
-    teamSize: pipelineData.jobId.teamSize,
-    numberOfPositions: pipelineData.jobId.numberOfPositions,
-    workVisa: pipelineData.jobId.workVisa ? 
-      (typeof pipelineData.jobId.workVisa === 'object' ? 
-        !!pipelineData.jobId.workVisa.workVisa : 
-        !!pipelineData.jobId.workVisa) : 
-      false,
-    gender: pipelineData.jobId.gender,
-    deadlineByClient: pipelineData.jobId.deadlineByClient || undefined,
-    keySkills: Array.isArray(pipelineData.jobId.specialization) ? 
-      pipelineData.jobId.specialization : 
-      [],
-    certifications: Array.isArray(pipelineData.jobId.certifications) ? 
-      pipelineData.jobId.certifications : 
-      [],
-    otherBenefits: Array.isArray(pipelineData.jobId.otherBenefits) ? 
-      pipelineData.jobId.otherBenefits.join(", ") : 
-      pipelineData.jobId.otherBenefits,
-    jobDescription: pipelineData.jobId.jobDescription,
-    // Client information from API
-    clientIndustry: pipelineData.jobId.client?.industry,
-    clientLocation: pipelineData.jobId.client?.location,
-    clientStage: pipelineData.jobId.client?.clientStage,
-    clientCountry: pipelineData.jobId.client?.countryOfBusiness,
-    clientWebsite: pipelineData.jobId.client?.website,
-    clientPhone: pipelineData.jobId.client?.phoneNumber,
-    clientEmails: pipelineData.jobId.client?.emails
-  };
-};
+// Local conversion utilities removed in favor of utils/convert
 
 export function RecruiterPipeline() {
   const { user } = useAuth();
